+++ ChildOfWeather/Data/Repository/Protocol/WeatherRepository.swift
@@ -1,0 +1,12 @@
+//
+//  WeatherRepository.swift
+//  ChildOfWeather
+//
+//  Created by 조영민 on 2022/04/05.
+//
+
+import Foundation
+
+protocol WeatherRepository {
+    
+}
