+++ ChildOfWeather/NetWork/Constant.swift
@@ -8,6 +8,7 @@
 import Foundation
 
 enum Constant {
+    
     static let key = "37bcd5b2997285a9b4a7fa952b140766"
     static let baseURL = "https://api.openweathermap.org/data/2.5/weather?"
 }
